+++ src/contract/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod apb_schema;
 pub(crate) mod atb_schema;
 pub(crate) mod axi_schema;
 pub(crate) mod common;